requirements.txt
python-docx==1.1.2
PyPDF2==3.0.1
pydantic==2.10.6
fastapi==0.115.7